aiBeTrade-stock/aibetrade_stock/events/pyproject.toml
[tool.poetry]
name = "aibetrade-stock"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.34.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
flask = "^3.0.3"
redis = "^5.0.3"
psycopg2-binary = "^2.9.9"
aiogram = "^3.7.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
